tests/submission/Infraware/Forms/contents/Forms/option_text.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>option_text</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype">
		  <option value="1"> Miner </option>
		  <option value="2"> Puffer </option>
		  <option value="3"> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
	</form>

	<script>
		
		var select = document.getElementById("unittype");
		var option = select.item(2);
		var text = option.text;
		option.text = 'CustomText';

		if (typeof(text) == "string") {
			if (text == 'Snipey' && option.text == 'CustomText') {
				test(function() {
					assert_true(true, "text attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("text attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("text attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
